% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGelman.R
\name{ggGelman}
\alias{ggGelman}
\title{Gelman Plotting function}
\usage{
ggGelman(list)
}
\arguments{
\item{list}{A list of Rhat values from JagsUI or saved from gelman.diag}
}
\description{
This function allows you to display Gelman diagonstics for
an entire JAGS model
}
\examples{
ggGelman(list = model$Rhat)

}
\keyword{convergence}
