% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGelman2.R
\name{ggGelman2}
\alias{ggGelman2}
\title{Per Parameter Gelman Plotting function}
\usage{
ggGelman2(list)
}
\arguments{
\item{list}{A list of Rhat values from JagsUI or saved from gelman.diag}
}
\description{
This function allows you to display Gelman diagonstics for
each parameter of choice in a JAGS model to find problem parameters
}
\examples{
ggGelman2()
}
\keyword{convergence}
