% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRaftery.R
\name{ggRaftery}
\alias{ggRaftery}
\title{Raftery diagnostics distribution function}
\usage{
ggRaftery(mcmc)
}
\arguments{
\item{mcmc}{MCMC samples from a JAGS run}
}
\description{
This function allows you to display a histogram of the distribution
of the number of iterations needed as suggested by the Raftery diagnostic
}
\examples{
#with Rjags
mcmc <- coda::mcmc(model)
ggRaftery(mcmc)

#with jagsUI
mcmc <- model$samples
ggRaftery(mcmc)

}
\keyword{convergence}
